+++ src/promotion/promotion.c
@@ -87,7 +87,7 @@ ocf_error_t ocf_promotion_set_policy(ocf_promotion_policy_t policy,
 }
 
 ocf_error_t ocf_promotion_set_param(ocf_promotion_policy_t policy,
-		uint8_t param_id, uint64_t param_value)
+		uint8_t param_id, uint32_t param_value)
 {
 	ocf_promotion_t type = policy->type;
 	ocf_error_t result = -OCF_ERR_INVAL;
@@ -103,7 +103,7 @@ ocf_error_t ocf_promotion_set_param(ocf_promotion_policy_t policy,
 }
 
 ocf_error_t ocf_promotion_get_param(ocf_promotion_policy_t policy,
-		uint8_t param_id, uint64_t *param_value)
+		uint8_t param_id, uint32_t *param_value)
 {
 	ocf_promotion_t type = policy->type;
 	ocf_error_t result = -OCF_ERR_INVAL;
